+++ server/pkg/integrationapi/csv_test.go
@@ -2,9 +2,7 @@ package integrationapi
 
 import (
 	"fmt"
-	"net/url"
 	"testing"
-	"time"
 
 	"github.com/reearth/reearth-cms/server/pkg/id"
 	"github.com/reearth/reearth-cms/server/pkg/item"
@@ -110,74 +108,8 @@ func TestToCSVProp(t *testing.T) {
 	assert.Equal(t, "Nour", s1)
 	assert.True(t, ok1)
 
-	sf2 := schema.NewField(schema.NewTextArea(lo.ToPtr(10)).TypeProperty()).NewID().Key(key.Random()).MustBuild()
-	if2 := item.NewField(sf2.ID(), value.TypeTextArea.Value("Nour").AsMultiple(), nil)
+	var if2 *item.Field
 	s2, ok2 := toCSVProp(if2)
-	assert.Equal(t, "Nour", s2)
-	assert.True(t, ok2)
-
-	sf3 := schema.NewField(schema.NewURL().TypeProperty()).NewID().Key(key.Random()).MustBuild()
-	v3 := url.URL{Scheme: "https", Host: "reearth.io"}
-	if3 := item.NewField(sf3.ID(), value.TypeURL.Value(v3).AsMultiple(), nil)
-	s3, ok3 := toCSVProp(if3)
-	assert.Equal(t, "https://reearth.io", s3)
-	assert.True(t, ok3)
-
-	sf4 := schema.NewField(schema.NewAsset().TypeProperty()).NewID().Key(key.Random()).MustBuild()
-	v4 := id.NewAssetID()
-	if4 := item.NewField(sf4.ID(), value.TypeAsset.Value(v4).AsMultiple(), nil)
-	s4, ok4 := toCSVProp(if4)
-	assert.Equal(t, v4.String(), s4)
-	assert.True(t, ok4)
-
-	// v5 := id.NewGroupID()
-	// sf5 := schema.NewField(schema.NewGroup(v5).TypeProperty()).NewID().Key(key.Random()).Multiple(true).MustBuild()
-	// if5 := item.NewField(sf5.ID(), value.MultipleFrom(value.TypeGroup, []*value.Value{value.TypeGroup.Value(v5)}), nil)
-	// s5, ok5 := toString(if5)
-	// assert.Equal(t, v5.String(), s5)
-	// assert.True(t, ok5)
-
-	v6 := id.NewItemID()
-	sf6 := schema.NewField(schema.NewReference(id.NewModelID(), id.NewSchemaID(), nil, nil).TypeProperty()).NewID().Key(key.Random()).MustBuild()
-	if6 := item.NewField(sf6.ID(), value.TypeReference.Value(v6).AsMultiple(), nil)
-	s6, ok6 := toCSVProp(if6)
-	assert.Equal(t, v6.String(), s6)
-	assert.True(t, ok6)
-
-	v7 := int64(30)
-	in7, _ := schema.NewInteger(lo.ToPtr(int64(1)), lo.ToPtr(int64(100)))
-	tp7 := in7.TypeProperty()
-	sf7 := schema.NewField(tp7).NewID().Name("age").Key(key.Random()).MustBuild()
-	if7 := item.NewField(sf7.ID(), value.TypeInteger.Value(v7).AsMultiple(), nil)
-	s7, ok7 := toCSVProp(if7)
-	assert.Equal(t, "30", s7)
-	assert.True(t, ok7)
-
-	v8 := float64(30.123)
-	in8, _ := schema.NewNumber(lo.ToPtr(float64(1)), lo.ToPtr(float64(100)))
-	tp8 := in8.TypeProperty()
-	sf8 := schema.NewField(tp8).NewID().Name("age").Key(key.Random()).MustBuild()
-	if8 := item.NewField(sf8.ID(), value.TypeNumber.Value(v8).AsMultiple(), nil)
-	s8, ok8 := toCSVProp(if8)
-	assert.Equal(t, "30.123", s8)
-	assert.True(t, ok8)
-
-	v9 := true
-	sf9 := schema.NewField(schema.NewBool().TypeProperty()).NewID().Name("age").Key(key.Random()).MustBuild()
-	if9 := item.NewField(sf9.ID(), value.TypeBool.Value(v9).AsMultiple(), nil)
-	s9, ok9 := toCSVProp(if9)
-	assert.Equal(t, "true", s9)
-	assert.True(t, ok9)
-
-	v10 := time.Now()
-	sf10 := schema.NewField(schema.NewDateTime().TypeProperty()).NewID().Name("age").Key(key.Random()).MustBuild()
-	if10 := item.NewField(sf10.ID(), value.TypeDateTime.Value(v10).AsMultiple(), nil)
-	s10, ok10 := toCSVProp(if10)
-	assert.Equal(t, v10.Format(time.RFC3339), s10)
-	assert.True(t, ok10)
-
-	var if11 *item.Field
-	s11, ok11 := toCSVProp(if11)
-	assert.Empty(t, s11)
-	assert.False(t, ok11)
+	assert.Empty(t, s2)
+	assert.False(t, ok2)
 }
